relocate the Rotate_Point function call and miscellaneous cleanup
